Make calculation double instead of float

+++ applications/spectrum_analyzer/spectrum_analyzer.c
@@ -135,9 +135,9 @@ static void spectrum_analyzer_render_callback(Canvas* const canvas, void* ctx) {
             temp_str,
             36,
             "Peak: %3.2f Mhz %3.1f dbm",
-            ((float)(model->channel0_frequency + (model->max_rssi_channel * model->spacing)) /
+            ((double)(model->channel0_frequency + (model->max_rssi_channel * model->spacing)) /
              1000000),
-            model->max_rssi);
+            (double) model->max_rssi);
         canvas_draw_str_aligned(canvas, 127, 0, AlignRight, AlignTop, temp_str);
     }
